requirements.txt
streamlit
pyaccsharedmemory
pandas
pywebview